+++ packages/widgetbook/lib/test.dart
@@ -1,1 +1,2 @@
+export 'src/test/guidelines/guidelines.dart';
 export 'src/test/test.dart' hide outputDir;